[IMP] tutorials/server_framework_101: rewrite the tutorial
The Server Framework 101 (formerly Getting Started) is generally seen as
an interesting and rewarding tutorial but also somewhat outdated and too
limited for beginners in Odoo development as it fails to introduce key
concepts of the server framework (e.g., controllers, tests, etc.). The
instructions are also too directive for the reader to try and search by
themselves and learn from it.

With this commit, all of the content of the tutorial is rewritten while
keeping the objective of building a real estate module. The setup guide
for tutorials is also improved to ensure smoother onboarding for Odoo
employees and community members alike.

task-3802536